Initial import of terraform.  This uses my keys and those need to be externalized.  Also, the ansible piece is not committed yet.
Full working example of terraform with ansible hooked in.

Terraform variables are externalized.  I need to test on a new machine
to verify no missing parts.

Before ansible can be run, a fix is there to create a symbolic link so
ansible will run on the AMI instance.  I tested several AMIs and none
had the necessary /usr/bin/python symbolic link.

After Terraform provisions the instance and the symbolic link is
created, Ansible is run to configure the instance.

After testing, I'll add instructions on how this can be set up on any
machine.

Adding the terraform.tfstate file so it can be shared among
provisionings.

Added the tfstate file to the .gitignore so it could be shared publically.
Added Amazon pem ssh files and anything else that should not be shared.
Added a README.md to explain how to set this code up for public sharing.

Added notes to inventories file.

Corrected notes on the site.yml ansible file.

Added a variable to be set for the ssh connection.  It is no longer hard
coded.

Cleaned up aws.tf, the terraform configuration file.

Externalized configuration variables.

Removed the tfstate file.

Removed the hard coded key variable names.